bithumb_LLM/requirements.txt
python-dotenv==1.0.1
openai==1.57.4
slack-bolt==1.18.1